Enum matrix_sdk_ui::room_list_service::RoomListLoadingState

pub enum RoomListLoadingState {
    NotLoaded,
    Loaded {
        maximum_number_of_rooms: Option<u32>,
    },
}

The loading state of a RoomList.

When a RoomList is displayed to the user, it can be in various states. This enum tries to represent those states with a correct level of abstraction.

Variants

NotLoaded

The RoomList has not been loaded yet, i.e. a sync might run or not run at all, there is nothing to show in this RoomList yet. It’s a good opportunity to show a placeholder to the user.

From Self::NotLoaded, it’s only possible to move to Self::Loaded.

Loaded

Fields

maximum_number_of_rooms: Option<u32>

The maximum number of rooms a RoomList contains.

It does not mean that there are exactly this many rooms to display. Usually, the room entries are represented by RoomListEntry. The room entry might have been synced or not synced yet, but we know for sure (from the server), that there will be this amount of rooms in the list at the end.

Note that it’s an Option, because it may be possible that the server did miss to send us this value. It’s up to you, dear reader, to know which default to adopt in case of None.

The RoomList has been loaded, i.e. a sync has been run, or more syncs are running, there is probably something to show to the user. Either the user has 0 room, in this case, it’s a good opportunity to show a special screen for that, or the user has multiple rooms, and it’s the classical room list.

The number of rooms is represented by maximum_number_of_rooms.

From Self::Loaded, it’s not possible to move back to Self::NotLoaded.