Module matrix_sdk_ui::timeline

A high-level view into a room’s contents.

See Timeline for details.

Modules

futures

Structs

EventTimelineItem

An item in the timeline that represents at least one event.

InReplyToDetails

Details about an event being replied to.

MemberProfileChange

An event changing a member’s profile.

Message

An m.room.message event or extensible event, including edits.

OtherState

A state event that doesn’t have its own variant.

PaginationOptions

Options for pagination.

PaginationOutcome

The result of a successful pagination request.

PollResult

Profile

The display name and avatar URL of a room member.

ReactionGroup

A group of reaction events on the same event with the same key.

ReactionSenderData

Data associated with a reaction sender. It can be used to display a details UI component for a reaction with both sender names and the date at which they sent a reaction.

RepliedToEvent

An event that is replied to.

RoomMembershipChange

An event changing a room membership.

Sticker

An m.sticker event.

Timeline

A high-level view into a regular¹ room’s contents.

TimelineBuilder

Builder that allows creating and configuring various parts of a Timeline.

TimelineItem

A single entry in timeline.

UnsupportedEditItem

UnsupportedReplyItem

Enums

AnyOtherFullStateEventContent

An enum over all the full state event contents that don’t have their own TimelineItemContent variant.

BackPaginationStatus

EncryptedMessage

Metadata about an m.room.encrypted event that could not be decrypted.

Error

Errors specific to the timeline.

EventItemOrigin

Where this event came.

EventSendState

This type represents the “send state” of a local event timeline item.

MembershipChange

An enum over all the possible room membership changes.

TimelineDetails

Some details of an EventTimelineItem that may require server requests other than just the regular sync_events.

TimelineItemContent

The content of an EventTimelineItem.

TimelineItemKind

VirtualTimelineItem

A TimelineItem that doesn’t correspond to an event.

Traits

RoomExt

SlidingSyncRoomExt

Functions

default_event_filter

The default event filter for crate::timeline::TimelineBuilder::event_filter.

Type Aliases

BundledReactions

The reactions grouped by key.