Struct matrix_sdk_ui::timeline::EventTimelineItem

pub struct EventTimelineItem { /* private fields */ }

An item in the timeline that represents at least one event.

There is always one main event that gives the EventTimelineItem its identity but in many cases, additional events like reactions and edits are also part of the item.

Implementations

impl EventTimelineItem

pub async fn from_latest_event( client: Client, room_id: &RoomId, latest_event: LatestEvent ) -> Option<EventTimelineItem>

If the supplied low-level SyncTimelineEventy is suitable for use as the latest_event in a message preview, wrap it as an EventTimelineItem.

pub fn is_local_echo(&self) -> bool

Check whether this item is a local echo.

This returns true for events created locally, until the server echoes back the full event as part of a sync response.

pub fn send_state(&self) -> Option<&EventSendState>

Get the event’s send state, if it is a local echo.

pub fn transaction_id(&self) -> Option<&TransactionId>

Get the transaction ID of this item.

The transaction ID is currently only kept until the remote echo for a local event is received.

pub fn event_id(&self) -> Option<&EventId>

Get the event ID of this item.

If this returns Some(_), the event was successfully created by the server.

Even if this is a local event, this can be Some(_) as the event ID can be known not just from the remote echo via sync_events, but also from the response of the send request that created the event.

pub fn sender(&self) -> &UserId

Get the sender of this item.

pub fn sender_profile(&self) -> &TimelineDetails<Profile>

Get the profile of the sender.

pub fn content(&self) -> &TimelineItemContent

Get the content of this item.

pub fn reactions(&self) -> &BundledReactions

Get the reactions of this item.

pub fn read_receipts(&self) -> &IndexMap<OwnedUserId, Receipt>

Get the read receipts of this item.

The key is the ID of a room member and the value are details about the read receipt.

Note that currently this ignores threads.

pub fn timestamp(&self) -> MilliSecondsSinceUnixEpoch

Get the timestamp of this item.

If this event hasn’t been echoed back by the server yet, returns the time the local event was created. Otherwise, returns the origin server timestamp.

pub fn is_own(&self) -> bool

Whether this timeline item was sent by the logged-in user themselves.

pub fn is_editable(&self) -> bool

Flag indicating this timeline item can be edited by current user.

pub fn is_highlighted(&self) -> bool

Whether the event should be highlighted in the timeline.

pub fn encryption_info(&self) -> Option<&EncryptionInfo>

Get the encryption information for the event, if any.

pub fn can_be_replied_to(&self) -> bool

Check whether this item can be replied to.

pub fn can_be_edited(&self) -> bool

Check whether this item can be edited.

Please also check whether the sender of this event is the client’s current user before presenting an edit button in the UI.

pub fn original_json(&self) -> Option<&Raw<AnySyncTimelineEvent>>

Get the raw JSON representation of the initial event (the one that caused this timeline item to be created).

Returns None if this event hasn’t been echoed back by the server yet.

pub fn latest_edit_json(&self) -> Option<&Raw<AnySyncTimelineEvent>>

Get the raw JSON representation of the latest edit, if any.

pub fn latest_json(&self) -> Option<&Raw<AnySyncTimelineEvent>>

Shorthand for item.latest_edit_json().or_else(|| item.original_json()).

pub fn origin(&self) -> Option<EventItemOrigin>

Get the origin of the event, i.e. where it came from.

May return None in some edge cases that are subject to change.