Struct matrix_sdk_ui::timeline::Timeline

pub struct Timeline { /* private fields */ }

A high-level view into a regular¹ room’s contents.

¹ This type is meant to be used in the context of rooms without a room_type, that is rooms that are primarily used to exchange text messages.

Implementations

impl Timeline

pub async fn clear(&self)

Clear all timeline items, and reset pagination parameters.

pub async fn paginate_backwards( &self, options: PaginationOptions<'_> ) -> Result<()>

Add more events to the start of the timeline.

pub async fn retry_decryption<'a, S: AsRef<str> + 'a>( &'a self, session_ids: impl IntoIterator<Item = &'a S> )

Retry decryption of previously un-decryptable events given a list of session IDs whose keys have been imported.

Examples

let path = PathBuf::from("/home/example/e2e-keys.txt");
let result =
    client.encryption().import_room_keys(path, "secret-passphrase").await?;

// Given a timeline for a specific room_id
if let Some(keys_for_users) = result.keys.get(room_id) {
    let session_ids = keys_for_users.values().flatten();
    timeline.retry_decryption(session_ids).await;
}

pub async fn item_by_event_id( &self, event_id: &EventId ) -> Option<EventTimelineItem>

Get the current timeline item for the given event ID, if any.

It’s preferable to store the timeline items in the model for your UI, if possible, instead of just storing IDs and coming back to the timeline object to look up items.

pub async fn items(&self) -> Vector<Arc<TimelineItem>>

Get the current list of timeline items. Do not use this in production!

pub async fn latest_event(&self) -> Option<EventTimelineItem>

Get the latest of the timeline’s event items.

pub async fn subscribe( &self ) -> (Vector<Arc<TimelineItem>>, impl Stream<Item = VectorDiff<Arc<TimelineItem>>>)

Get the current timeline items, and a stream of changes.

You can poll this stream to receive updates. See futures_util::StreamExt for a high-level API on top of Stream.

pub async fn subscribe_filter_map<U: Clone>( &self, f: impl Fn(Arc<TimelineItem>) -> Option<U> ) -> (Vector<U>, impl Stream<Item = VectorDiff<U>>)

pub async fn send( &self, content: AnyMessageLikeEventContent, txn_id: Option<&TransactionId> )

Send a message to the room, and add it to the timeline as a local echo.

For simplicity, this method doesn’t currently allow custom message types.

If the encryption feature is enabled, this method will transparently encrypt the room message if the room is encrypted.

If sending the message fails, the local echo item will change its send_state to EventSendState::SendingFailed.

Arguments

content - The content of the message event.

txn_id - A locally-unique ID describing a message transaction with the homeserver. Unless you’re doing something special, you can pass in None which will create a suitable one for you automatically.

On the sending side, this field is used for re-trying earlier failed transactions. Subsequent messages must never re-use an earlier transaction ID.

On the receiving side, the field is used for recognizing our own messages when they arrive down the sync: the server includes the ID in the MessageLikeUnsigned field transaction_id of the corresponding SyncMessageLikeEvent, but only for the sending device. Other devices will not see it.

pub async fn toggle_reaction( &self, annotation: &Annotation ) -> Result<(), Error>

Toggle a reaction on an event

Adds or redacts a reaction based on the state of the reaction at the time it is called.

When redacting an event, the redaction reason is not sent.

Ensures that only one reaction is sent at a time to avoid race conditions and spamming the homeserver with requests.

pub fn send_attachment( &self, url: String, mime_type: Mime, config: AttachmentConfig ) -> SendAttachment<'_>

Sends an attachment to the room. It does not currently support local echoes

If the encryption feature is enabled, this method will transparently encrypt the room message if the room is encrypted.

Arguments

url - The url for the file to be sent

mime_type - The attachment’s mime type

config - An attachment configuration object containing details about the attachment like a thumbnail, its size, duration etc.

pub async fn retry_send(&self, txn_id: &TransactionId) -> Result<(), Error>

Retry sending a message that previously failed to send.

Arguments

txn_id - The transaction ID of a local echo timeline item that has a send_state() of SendState::FailedToSend { .. }

pub async fn cancel_send(&self, txn_id: &TransactionId) -> bool

Discard a local echo for a message that failed to send.

Returns whether the local echo with the given transaction ID was found.

Argument

txn_id - The transaction ID of a local echo timeline item that has a send_state() of SendState::FailedToSend { .. }. Note: A send state of SendState::NotYetSent might be supported in the future as well, but there can be no guarantee for that actually stopping the event from reaching the server.

pub async fn fetch_details_for_event( &self, event_id: &EventId ) -> Result<(), Error>

Fetch unavailable details about the event with the given ID.

This method only works for IDs of remote EventTimelineItems, to prevent losing details when a local echo is replaced by its remote echo.

This method tries to make all the requests it can. If an error is encountered for a given request, it is forwarded with the TimelineDetails::Error variant.

Arguments

event_id - The event ID of the event to fetch details for.

Errors

Returns an error if the identifier doesn’t match any event with a remote echo in the timeline, or if the event is removed from the timeline before all requests are handled.

pub async fn fetch_members(&self)

Fetch all member events for the room this timeline is displaying.

If the full member list is not known, sender profiles are currently likely not going to be available. This will be fixed in the future.

If fetching the members fails, any affected timeline items will have the sender_profile set to TimelineDetails::Error.

pub async fn latest_user_read_receipt( &self, user_id: &UserId ) -> Option<(OwnedEventId, Receipt)>

Get the latest read receipt for the given user.

Contrary to Common::user_receipt() that only keeps track of read receipts received from the homeserver, this keeps also track of implicit read receipts in this timeline, i.e. when a room member sends an event.

pub async fn send_single_receipt( &self, receipt_type: ReceiptType, thread: ReceiptThread, event_id: OwnedEventId ) -> Result<()>

Send the given receipt.

This uses Joined::send_single_receipt internally, but checks first if the receipt points to an event in this timeline that is more recent than the current ones, to avoid unnecessary requests.

pub async fn send_multiple_receipts(&self, receipts: Receipts) -> Result<()>

Send the given receipts.

This uses Joined::send_multiple_receipts internally, but checks first if the receipts point to events in this timeline that are more recent than the current ones, to avoid unnecessary requests.