Trait matrix_sdk_ui::timeline::SlidingSyncRoomExt

pub trait SlidingSyncRoomExt {
    // Required methods
    fn timeline<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<Timeline>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn latest_timeline_item<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = Option<EventTimelineItem>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Required Methods

fn timeline<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Option<Timeline>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get a Timeline for this room.

fn latest_timeline_item<'life0, 'async_trait>( &'life0 self ) -> Pin<Box<dyn Future<Output = Option<EventTimelineItem>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Get the latest timeline item of this room, if it is already cached.

Use Timeline::latest_event instead if you already have a timeline for this SlidingSyncRoom.